more refractoring done to eliminate duplication
